Moved MessageConsumer into mod.rs in iggy_consumer_ext
