+++ sdk/src/iggy_consumer_ext/iggy_consumer_message_trait.rs
@@ -1,8 +1,8 @@
 use async_trait::async_trait;
 
 use crate::error::IggyError;
-use crate::message_consumer::MessageConsumer;
 use tokio::sync::oneshot;
+use crate::iggy_consumer_ext::MessageConsumer;
 
 #[async_trait]
 pub trait IggyConsumerMessageExt {
